Renaming HAL/Conversion/IREEToHAL -> UtilToHAL.
