+++ iree/compiler/Dialect/HAL/Conversion/UtilToHAL/ConvertUtilToHAL.cpp
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/compiler/Dialect/HAL/Conversion/IREEToHAL/ConvertIREEToHAL.h"
+#include "iree/compiler/Dialect/HAL/Conversion/UtilToHAL/ConvertUtilToHAL.h"
 
 #include "iree/compiler/Dialect/HAL/IR/HALOps.h"
 #include "iree/compiler/Dialect/Util/IR/UtilOps.h"
@@ -75,64 +75,13 @@ class DynamicShapeConstantOpConversion
   }
 };
 
-template <typename T>
-class GenericConvertTypesConversion : public OpConversionPattern<T> {
- public:
-  using OpConversionPattern<T>::OpConversionPattern;
-
-  LogicalResult matchAndRewrite(
-      T op, llvm::ArrayRef<Value> newOperands,
-      ConversionPatternRewriter &rewriter) const override {
-    SmallVector<Type> newTypes;
-    bool anyChanged = false;
-    for (auto oldNew : llvm::zip(op->getOperands(), newOperands)) {
-      auto oldValue = std::get<0>(oldNew);
-      auto newValue = std::get<1>(oldNew);
-      if (oldValue.getType() != newValue.getType()) {
-        anyChanged = true;
-        break;
-      }
-    }
-    for (auto oldType : op.getOperation()->getResultTypes()) {
-      auto newType = this->getTypeConverter()->convertType(oldType);
-      if (oldType != newType) anyChanged = true;
-      newTypes.push_back(newType);
-    }
-    if (!anyChanged) return failure();
-    rewriter.replaceOpWithNewOp<T>(op, newTypes, newOperands, op->getAttrs());
-    return success();
-  }
-};
-
 }  // namespace
 
-void populateIREEToHALPatterns(MLIRContext *context, ConversionTarget &target,
+void populateUtilToHALPatterns(MLIRContext *context, ConversionTarget &target,
                                TypeConverter &typeConverter,
                                OwningRewritePatternList &patterns) {
   target.addIllegalOp<IREE::Util::DynamicShapeConstantOp>();
   patterns.insert<DynamicShapeConstantOpConversion>(context);
-
-  typeConverter.addConversion([&](IREE::Util::ListType type) {
-    auto elementType = typeConverter.convertType(type.getElementType());
-    return IREE::Util::ListType::get(elementType);
-  });
-
-  target.addDynamicallyLegalOp<IREE::Util::ListCreateOp>(
-      [&](IREE::Util::ListCreateOp op) {
-        return typeConverter.isLegal(op.getType());
-      });
-  target.addDynamicallyLegalOp<IREE::Util::ListGetOp>(
-      [&](IREE::Util::ListGetOp op) {
-        return typeConverter.isLegal(op.getType());
-      });
-  target.addDynamicallyLegalOp<IREE::Util::ListSetOp>(
-      [&](IREE::Util::ListSetOp op) {
-        return typeConverter.isLegal(op.value().getType());
-      });
-  patterns.insert<GenericConvertTypesConversion<IREE::Util::ListCreateOp>,
-                  GenericConvertTypesConversion<IREE::Util::ListGetOp>,
-                  GenericConvertTypesConversion<IREE::Util::ListSetOp>>(
-      typeConverter, context);
 }
 
 }  // namespace iree_compiler
