+++ iree/compiler/Dialect/HAL/Conversion/UtilToHAL/test/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# iree/compiler/Dialect/HAL/Conversion/IREEToHAL/BUILD                         #
+# iree/compiler/Dialect/HAL/Conversion/UtilToHAL/test/BUILD                    #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -10,20 +10,14 @@
 
 iree_add_all_subdirs()
 
-iree_cc_library(
+iree_lit_test_suite(
   NAME
-    IREEToHAL
-  HDRS
-    "ConvertIREEToHAL.h"
+    lit
   SRCS
-    "ConvertIREEToHAL.cpp"
-  DEPS
-    MLIRIR
-    MLIRStandard
-    MLIRTransforms
-    iree::compiler::Dialect::HAL::IR
-    iree::compiler::Dialect::Util::IR
-  PUBLIC
+    "shape_constants.mlir"
+  DATA
+    iree::tools::IreeFileCheck
+    iree::tools::iree-opt
 )
 
 ### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
