+++ iree/compiler/Dialect/HAL/Transforms/BUILD
@@ -40,8 +40,8 @@ cc_library(
         "//iree/compiler/Dialect/HAL/Conversion",
         "//iree/compiler/Dialect/HAL/Conversion/FlowToHAL",
         "//iree/compiler/Dialect/HAL/Conversion/HALToHAL",
-        "//iree/compiler/Dialect/HAL/Conversion/IREEToHAL",
         "//iree/compiler/Dialect/HAL/Conversion/StandardToHAL",
+        "//iree/compiler/Dialect/HAL/Conversion/UtilToHAL",
         "//iree/compiler/Dialect/HAL/IR",
         "//iree/compiler/Dialect/HAL/IR:HALDialect",
         "//iree/compiler/Dialect/HAL/Target",
