+++ iree/compiler/Dialect/VM/Conversion/UtilToVM/ConvertUtilToVM.cpp
@@ -25,6 +25,7 @@ void populateUtilGlobalToVMPatterns(MLIRContext *context,
                                     TypeConverter &typeConverter,
                                     OwningRewritePatternList &patterns);
 void populateUtilListToVMPatterns(MLIRContext *context,
+                                  ConversionTarget &conversionTarget,
                                   TypeConverter &typeConverter,
                                   OwningRewritePatternList &patterns);
 void populateUtilStatusToVMPatterns(MLIRContext *context,
@@ -114,7 +115,8 @@ void populateUtilToVMPatterns(MLIRContext *context,
 
   populateUtilGlobalToVMPatterns(context, conversionTarget, typeConverter,
                                  patterns);
-  populateUtilListToVMPatterns(context, typeConverter, patterns);
+  populateUtilListToVMPatterns(context, conversionTarget, typeConverter,
+                               patterns);
   populateUtilStatusToVMPatterns(context, conversionTarget, typeConverter,
                                  patterns);
 }
